work on correct mouse input
