+++ compile_commands.json
@@ -6,6 +6,16 @@
 },
 {
   "directory": "/mnt/c/Users/User/Projects/CG-labs/build",
+  "command": "/usr/bin/c++  -I/mnt/c/Users/User/Projects/CG-labs/include -g -std=gnu++20 -o CMakeFiles/src_lib.dir/src/EventHandlers/LookupHandler.cpp.o -c /mnt/c/Users/User/Projects/CG-labs/src/EventHandlers/LookupHandler.cpp",
+  "file": "/mnt/c/Users/User/Projects/CG-labs/src/EventHandlers/LookupHandler.cpp"
+},
+{
+  "directory": "/mnt/c/Users/User/Projects/CG-labs/build",
+  "command": "/usr/bin/c++  -I/mnt/c/Users/User/Projects/CG-labs/include -g -std=gnu++20 -o CMakeFiles/src_lib.dir/src/EventHandlers/moveHandler.cpp.o -c /mnt/c/Users/User/Projects/CG-labs/src/EventHandlers/moveHandler.cpp",
+  "file": "/mnt/c/Users/User/Projects/CG-labs/src/EventHandlers/moveHandler.cpp"
+},
+{
+  "directory": "/mnt/c/Users/User/Projects/CG-labs/build",
   "command": "/usr/bin/c++  -I/mnt/c/Users/User/Projects/CG-labs/include -g -std=gnu++20 -o CMakeFiles/src_lib.dir/src/Light/Lambert.cpp.o -c /mnt/c/Users/User/Projects/CG-labs/src/Light/Lambert.cpp",
   "file": "/mnt/c/Users/User/Projects/CG-labs/src/Light/Lambert.cpp"
 },
