+++ include/Scene.hpp
@@ -2,6 +2,7 @@
 #define SCENE_HPP_
 
 #include "CameraMVP.hpp"
+#include "Light/Light.hpp"
 #include "Program.hpp"
 #include "objects/Object.hpp"
 #include <list>
@@ -10,9 +11,10 @@ class Scene {
   std::list<Object*> objects_;
   CameraMVP cameraData_;
   Program & program_;
+  ILight & light_;
 
 public:
-  Scene(Program & program, CameraMVP cameraData, std::list<Object*> objects = {});
+  Scene(Program & program, CameraMVP cameraData, ILight & light, std::list<Object*> objects = {});
 
   void update(double time, double dt);
 };
