+++ include/glCheckError.hpp
@@ -1,14 +1,16 @@
 #ifndef GL_CHECK_ERROR_HPP_
 #define GL_CHECK_ERROR_HPP_
 
-#include <string>
-#include <iostream>
 #define GLEW_STATIC
 #include <GL/glew.h>
 
 
 GLenum glCheckError_(const char *file, int line);
+#ifdef DEBUG
 #define glCheckError() glCheckError_(__FILE__, __LINE__)
+#else
+#define glCheckError() {};
+#endif
 
 void glDebugOutput(GLenum source, 
                             GLenum type,
